+++ web/app/store/products/reducer.js
@@ -1,15 +1,14 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receivePlpProductData} from './actions'
-
 import {receivePdpProductData} from '../../integration-manager/responses'
+import {receiveProductListProductData} from './actions'
 
 const initialState = Immutable.Map()
 
 const productReducer = handleActions({
     ...mergePayloadForActions(receivePdpProductData),
-    [receivePlpProductData]: (state, {payload}) => {
+    [receiveProductListProductData]: (state, {payload}) => {
         return state.mergeDeepWith((prev, next) => prev || next, payload)
     }
 }, initialState)