+++ web/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## To be released
 - Add option to analyze bundle after build. Set `MOBIFY_ANALYZE` environment variable to `true` before running any build. Eg. `MOBIFY_ANALYZE=true npm run prod:build`
+- Rename plp and pdp containers to product-list and product-details
+- Rename product-list component in checkout to order-summary
 
 ## 0.11.0 (February 10, 2017)
 - Upgrade to the latest SDK